pyproject.toml
[project]
name = "ic"
version = "0.1.0"
description = ""
authors = [
    {name = "Gabriel Azevedo",email = "g.azevedo0121@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "minio (>=7.2.16,<8.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "ipython (>=9.4.0,<10.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic (>=2.11.7,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
